IndexedDB/idbobjectstore_count.htm

<!DOCTYPE html>
<html>
<head>
    <meta content="text/html; charset=utf-8" http-equiv="content-type" />
    <title> IDBObjectStore.count() - returns the number of records in the object store </title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript">
        var objectStoreName = "objectstore";
        var db = null;
        
        var t = async_test();
        
        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName, databaseVersion);
            rqOpen.onupgradeneeded = t.step_func(function(event) {
                db = event.target.result;
                db.onerror = t.step_func(assert_database_error);
                var objStore = db.createObjectStore(objectStoreName);
                for(var i = 0; i < 10; i++) {
                    objStore.add({property:"data"}, i);
                }
                var rqCount = objStore.count();
                rqCount.onsuccess = t.step_func(function(event) {
                    assert_equals(event.target.result, 10, 'event.target.result');
                    t.done();
                });
            });
            rqOpen.onerror = t.step_func(assert_open_request_error);
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
